requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.5.0=h63175ca_0
blas=1.0=mkl
box2d-py=2.3.8=py39h99910a6_7
brotli=1.0.9=hcfcfb64_8
brotli-bin=1.0.9=hcfcfb64_8
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.12.7=h5b45459_0
cairo=1.16.0=hdecc03f_1015
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py39h68f70e3_3
cloudpickle=2.2.1=pyhd8ed1ab_0
contourpy=1.0.7=py39h1f6ef14_0
cycler=0.11.0=pyhd8ed1ab_0
dav1d=1.0.0=h8ffe710_1
dill=0.3.6=pyhd8ed1ab_1
expat=2.5.0=h1537add_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=hbde0cde_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.39.3=py39ha55989b_0
freetype=2.12.1=h546665d_1
future=0.18.3=pyhd8ed1ab_0
gettext=0.21.1=h5728263_0
graphite2=1.3.13=1000
gymnasium=0.27.1=py39h00e5391_0
gymnasium-box2d=0.27.1=py39hcbf5309_0
gymnasium-classic_control=0.27.1=py39hcbf5309_0
gymnasium-notices=0.0.1=pyh1a96a4e_0
harfbuzz=6.0.0=h196d34a_1
icu=72.1=h63175ca_0
importlib-metadata=6.1.0=pyha770c72_0
importlib-resources=5.12.0=pyhd8ed1ab_0
importlib_metadata=6.1.0=hd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2023.0.0=h57928b3_25922
jax-jumpy=1.0.0=pyhd8ed1ab_0
jpeg=9e=hcfcfb64_3
kiwisolver=1.4.4=py39h1f6ef14_1
lcms2=2.15=ha5c8aab_0
lerc=4.0.0=h63175ca_0
libavif=0.11.1=h8f5c5bc_0
libblas=3.9.0=12_win64_mkl
libbrotlicommon=1.0.9=hcfcfb64_8
libbrotlidec=1.0.9=hcfcfb64_8
libbrotlienc=1.0.9=hcfcfb64_8
libcblas=3.9.0=12_win64_mkl
libdeflate=1.17=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libflac=1.4.2=h63175ca_0
libglib=2.74.1=he8f3873_1
libhwloc=2.9.0=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=12_win64_mkl
libmad=0.15.1b=hcfcfb64_1001
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libtiff=4.5.0=hf8721a0_2
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=h0e60522_0
libwebp-base=1.3.0=hcfcfb64_0
libxcb=1.13=hcd874cb_1004
libxml2=2.10.3=hc3477c8_6
libzlib=1.2.13=hcfcfb64_4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib-base=3.7.1=py39haf65ace_0
mkl=2021.4.0=h0e2418a_729
mkl-service=2.4.0=py39h6b0492b_0
mpg123=1.31.3=h63175ca_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=pyh9f0ad1d_0
ninja=1.11.1=h91493d7_0
numpy=1.24.2=py39h16ffa76_0
openjpeg=2.5.0=ha2aaf27_2
openssl=1.1.1t=hcfcfb64_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=py39h2ba5b7c_1
pcre2=10.40=h17e33f8_0
pillow=9.4.0=py39hcebd2be_1
pip=23.0.1=py39haa95532_0
pixman=0.40.0=h8ffe710_0
portmidi=2.0.4=h63175ca_0
pthread-stubs=0.4=hcd874cb_1001
pthreads-win32=2.9.1=hfa6e2cd_3
pycparser=2.21=pyhd8ed1ab_0
pygad=2.19.1=pyhd8ed1ab_0
pygame=2.1.3=py39h928c5e4_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.9.16=h6244533_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.12.1=cpu_py39h5e1f01c_0
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py39ha55989b_5
sdl2=2.26.4=h63175ca_0
sdl2_image=2.6.3=h090e09f_0
sdl2_mixer=2.6.3=h090ff69_0
sdl2_ttf=2.20.2=h41be9cf_0
setuptools=65.6.3=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.1=h2bbff1b_0
tbb=2021.8.0=h91493d7_0
tk=8.6.12=h8ffe710_0
typing-extensions=4.5.0=hd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.0.0=py39ha55989b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.34.31931=h4c5c07a_10
wheel=0.38.4=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=h8ffe710_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h12be248_6
